+++ efpmd/torch/ANIModel.h
@@ -14,11 +14,9 @@ public:
     ~ANIModel() {printf("\nANIModel() Destructor\n");}
     void load_model(int model_type, const std::string &nn_path);
     void load_custom_model(const std::string &aev_name, const std::string &model_name, const std::string &nn_path);
-    //void get_energy_grad(const torch::Tensor& coordinates, const torch::Tensor& species, float* atomic_energies, float* gradients, float* forces, int num_atoms);
-    void get_energy_grad(const torch::Tensor& coordinates, const torch::Tensor& species, double* total_energy, float* gradients, float* forces, int num_atoms, int print); 
-    void get_custom_energy_grad(float* coordinates_data, int64_t* species_data, float* elecpots_data, int num_atoms, double* custom_energy, float* cus_grads, float* cus_forces, int print);
-    //void get_custom_energy_grad(float* coordinates_data, int64_t* species_data, float* elecpots_data, int num_atoms, float* custom_energy, float* cus_grads, float* cus_forces);
-private:
+    void get_energy_grad(const torch::Tensor& coordinates, const torch::Tensor& species, double* total_energy, float* gradients, float* forces, int num_atoms, int print);
+    void get_custom_energy_grad(double* coordinates_data, int64_t* species_data, double* elecpots_data, int num_atoms, double* custom_energy, double* cus_grads, double* cus_forces, int print);
+private: 
     torch::jit::Module module;
     torch::jit::Module aev_computer;
 };
